deletePost mutation authorization done

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -32,23 +32,28 @@ const Mutation = {
       user, token: jwt.sign({userId: user.id}, 'thisismysecret')
     }
   },
-  async deleteUser(parent, args, { prisma }, info) {
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.deleteUser({
       where: {
-        id: args.id
+        id: userId
       }
     }, info);
   },
-  async updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.updateUser({
       where: {
-        id: args.id
+        id: userId
       },
       data: args.data
     }, info);
   },
   async createPost(parent, args, { prisma, request }, info) {
     const userId = getUserId(request);
+
     return prisma.mutation.createPost({
       data: {
         title: args.data.title,
@@ -63,7 +68,17 @@ const Mutation = {
       
     }, info);
   },
-  async deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if(!postExists) throw new Error('Unable to delete post')
+
     return prisma.mutation.deletePost({
       where: {
         id: args.id
